games refactor

+++ BoardGameTracker.Common/Models/GameStatistics.cs
@@ -4,7 +4,7 @@ public class GameStatistics
 {
     public int PlayCount { get; set; }
     public double TotalPlayedTime { get; set; }
-    public double? PricePerPlay { get; set; }
+    public decimal? PricePerPlay { get; set; }
     public double? HighScore { get; set; }
     public double? AverageScore { get; set; }
     public MostWinningPlayer? MostWinsPlayer { get; set; }
